Remove former installation
