+++ pulse/core/core_group.py
@@ -4,9 +4,6 @@ from pulse.config.config_configure import configure
 from pulse.init.init import init
 from pulse.pods.pods import pods
 from pulse.build.build import build
-from pulse.package.package_install import install
-from pulse.package.package_uninstall import uninstall
-from pulse.package.package_ensure import ensure
 from pulse.run.run import run
 from pulse.stroke.stroke import stroke
 from pulse.package.package_pack import package
@@ -37,9 +34,6 @@ pulse.add_command(init)
 pulse.add_command(configure)
 pulse.add_command(pods)
 pulse.add_command(build)
-pulse.add_command(install)
-pulse.add_command(uninstall)
-pulse.add_command(ensure)
 pulse.add_command(run)
 pulse.add_command(stroke)
 pulse.add_command(package)
