--- pulse/package/content/__init__.py
@@ -1,15 +1,0 @@
-from .package_content import (
-    get_requirements,
-    get_package_syntax,
-    get_package_type,
-    get_local_package_type,
-    echo_retrieve_fail
-)
-
-from .package_resource_content import (
-    get_package_resources,
-    get_resource_files,
-    get_resource_includes,
-    get_resource_plugins,
-    get_resource_repo
-)
